Remove ResponsiveBackground component
Support for responsive images in browsers is limited, you cannot have all the possibilities of the browser picking the best supported file format *and* size, as `<picture>` and `srcset` combined can do. With the `object-fit` CSS property you can still create a `<img>` that behaves like a background. For now this is left to the user to implement, as we need to concentrate on getting the next-gen version of plain images working.