Delay setting src attribute after <img> is in DOM
This is an interesting problem: some browsers (FF, Safari, but not Chrome) would load an image immediately one you set the `src` attribute of `<img>`. While this is expected (think of programmatically preloading images), it poses a problem when you actually want to use the `<img>` as a child of `<picture>`, as the `src` attribute represents the fallback for browsers not supporting `<source>`.

But this is actually what happens if you render the markup in Ember: it creates the element detached from DOM (`document.createElement()`), sets the attributes (`element.setAttribute()`) and only after that attaches it to its parent. Chrome handles this more cleverly, but FF would (upon setting `src`) load the fallback image (jpeg), later sees the better `<source>` alternative (e.g. webp) and would load that *also*. Note that this does *not* happen, when the `<img>` is already part of `<picture>`, as the browser probably can understand now that the `src` attribute is not the only possible image URL.

Also all of that does not happen in a server-rendered HTML page (FastBoot, or any other server-rendered page), only when JS/Ember re-renders the page with JavaScript would that wrong behaviour show up again.

The solution seems simple enough: delay setting `src` until the `<img>` is in DOM and part of `<picture>`.